Some methods added and coverage brought to 100%

+++ src/main/java/it/attsd/deepsky/service/ConstellationService.java
@@ -3,6 +3,9 @@ package it.attsd.deepsky.service;
 import java.util.List;
 
 import it.attsd.deepsky.exceptions.ConstellationAlreadyExistsException;
+import it.attsd.deepsky.exceptions.ConstellationIsStillUsedException;
+import it.attsd.deepsky.model.DeepSkyObject;
+import it.attsd.deepsky.repository.DeepSkyObjectRepository;
 import org.springframework.stereotype.Service;
 
 import it.attsd.deepsky.model.Constellation;
@@ -11,40 +14,51 @@ import it.attsd.deepsky.repository.ConstellationRepository;
 @Service
 public class ConstellationService {
 
-	private ConstellationRepository constellationRepository;
-	
-	public ConstellationService(ConstellationRepository constellationRepository) {
-		this.constellationRepository = constellationRepository;
-	}
+    private ConstellationRepository constellationRepository;
 
-	public List<Constellation> findAll() {
-		return constellationRepository.findAll();
-	}
+    private DeepSkyObjectRepository deepSkyObjectRepository;
 
-	public Constellation findById(long id) {
-		return constellationRepository.findById(id).orElse(null);
-	}
+    public ConstellationService(ConstellationRepository constellationRepository, DeepSkyObjectRepository deepSkyObjectRepository) {
+        this.constellationRepository = constellationRepository;
+        this.deepSkyObjectRepository = deepSkyObjectRepository;
+    }
 
-	public Constellation findByName(String name) {
-		return constellationRepository.findByName(name);
-	}
+    public List<Constellation> findAll() {
+        return constellationRepository.findAll();
+    }
 
-	public Constellation save(Constellation constellation) throws ConstellationAlreadyExistsException {
-		Constellation existingConstellation = constellationRepository.findByName(constellation.getName());
-		if(existingConstellation != null){
-			throw new ConstellationAlreadyExistsException();
-		}
-		constellation.setId(null);
-		return constellationRepository.save(constellation);
-	}
+    public Constellation findById(long id) {
+        return constellationRepository.findById(id).orElse(null);
+    }
 
-	public Constellation updateById(long id, Constellation constellation) {
-		constellation.setId(id);
-		return constellationRepository.save(constellation);
-	}
+    public Constellation findByName(String name) {
+        return constellationRepository.findByName(name);
+    }
 
-	public void deleteById(long id) {
-		constellationRepository.deleteById(id);
-	}
+    public Constellation save(Constellation constellation) throws ConstellationAlreadyExistsException {
+        Constellation existingConstellation = constellationRepository.findByName(constellation.getName());
+        if (existingConstellation != null) {
+            throw new ConstellationAlreadyExistsException();
+        }
+        constellation.setId(null);
+        return constellationRepository.save(constellation);
+    }
+
+    public Constellation updateById(long id, Constellation constellation) {
+        constellation.setId(id);
+        return constellationRepository.save(constellation);
+    }
+
+    public void deleteById(long id) throws ConstellationIsStillUsedException {
+        Constellation existingConstellation = constellationRepository.findById(id).orElse(null);
+        if (existingConstellation != null) {
+            List<DeepSkyObject> deepSkyObjects = deepSkyObjectRepository.findByConstellation(existingConstellation);
+            if (!deepSkyObjects.isEmpty()) {
+                throw new ConstellationIsStillUsedException();
+            }
+            constellationRepository.deleteById(id);
+        }
+
+    }
 
 }
